# GardenControlUI/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "GardenControlApi",
    "description": "An API that allows the monitor and control of devices plugged in to the GPIO Pins on a Raspberry Pi",
    "version": "v1"
  },
  "paths": {
    "/api/AppSettings": {
      "get": {
        "tags": [
          "AppSettings"
        ],
        "summary": "Returns all settings",
        "responses": {
          "200": {
            "description": "Returns the all AppSettings",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppSettingDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppSettingDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppSettingDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AppSettings"
        ],
        "summary": "Creates an AppSetting",
        "description": "Any value set for CanBeUpdated property will be ignored and always set to true",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppSettingDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AppSettingDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AppSettingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Returns the newly created AppSetting",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AppSettingDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppSettingDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppSettingDto"
                }
              }
            }
          },
          "400": {
            "description": "If the key or value is null or empty",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/AppSettings/{id}": {
      "get": {
        "tags": [
          "AppSettings"
        ],
        "summary": "Returns specified setting",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the the specified AppSetting",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AppSettingDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppSettingDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppSettingDto"
                }
              }
            }
          },
          "404": {
            "description": "If the AppSetting cannot be found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "AppSettings"
        ],
        "summary": "Updates setting with specified value",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppSettingDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AppSettingDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AppSettingDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Returns when the AppSetting is updated sucessfully"
          },
          "400": {
            "description": "If the id parameter and id in the object do not match",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "If the AppSetting cannot be found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "AppSettings"
        ],
        "summary": "Deletes specified setting",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Returns when the AppSetting is deleted sucessfully"
          },
          "404": {
            "description": "If the AppSetting cannot be found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AppSettingDto": {
        "required": [
          "appSettingId",
          "key",
          "value"
        ],
        "type": "object",
        "properties": {
          "appSettingId": {
            "type": "integer",
            "format": "int32"
          },
          "key": {
            "maxLength": 255,
            "type": "string"
          },
          "value": {
            "type": "string"
          },
          "canBeUpdated": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  }
}